Move reminders controller to additional payments journey namespace
