+++ app/views/additional_payments/reminders/_one_time_password.html.erb
