+++ app/views/additional_payments/reminders/email_verification.html.erb
@@ -1,0 +1,3 @@
+<% content_for(:page_title, page_title(t("one_time_password.title"), journey: current_journey_routing_name, show_error: current_reminder.errors.any?)) %>
+
+<%= render partial: "reminders/one_time_password", locals: {current_reminder: current_reminder, current_journey_routing_name: current_journey_routing_name} %>
