+++ app/views/additional_payments/reminders/personal_details.html.erb
